+++ admin/components/Controller.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * @link http://www.56hm.com/repar
+ * @copyright Copyright (c) 2014 56hm LLC
+ * @license http://www.56hm.com/repar/license
+ */
+
+namespace admin\components;
+
+
+/**
+ * backend base Controller
+ *
+ * @author Repar <47558328@qq.com>
+ * @since  1.0
+ */
+abstract class Controller extends \common\components\Controller {
+
+}