+++ admin/controllers/DefaultController.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * @link http://www.56hm.com/repar
+ * @copyright Copyright (c) 2014 56hm LLC
+ * @license http://www.56hm.com/repar/license
+ */
+
+namespace admin\controllers;
+
+
+use admin\components\Controller;
+/**
+ * 
+ *
+ * @author Repar <47558328@qq.com>
+ * @since  1.0
+ */
+class DefaultController extends Controller{
+
+    public function actionIndex(){
+    	
+    }
+}