+++ common/components/Controller.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * @link http://www.56hm.com/repar
+ * @copyright Copyright (c) 2014 56hm LLC
+ * @license http://www.56hm.com/repar/license
+ */
+
+namespace common\components;
+
+
+/**
+ * base Controller
+ * 
+ * @author Repar <47558328@qq.com>
+ * @since  1.0
+ */
+class Controller extends \yii\web\Controller{
+
+
+}