+++ common/config/aliases.php
@@ -1,0 +1,6 @@
+<?php
+return [
+
+    'pack' => dirname(__DIR__) . '/pack';
+
+];