+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 ## Changelog
 
 
+**v0.2.0**
+
+release: 14 Nov 2014
+
+- added base Controllers
+- added public resource package 'pack' directory
+- added aliases config file
+
+
+
+
 **v0.1.0**
 
 release: 14 Nov 2014